requirements.txt
Flask==2.2.2
psycopg2-binary==2.9.3
Flask-SQLAlchemy==3.0.2
Flask-Migrate==3.1.0
pytest==7.1.3
google-cloud-storage==2.1.0
python-dotenv==0.21.0
flask-restx==1.0.3